deleted commented code and formatted
Issue that took so long to fix in the last commit was
due to the fact that I forgot setState causes a rerendering
to occur. So with the setTimer and setInterval both calling
setState, we were rerendering a lot! So I just made the
setInterval time value random per component, so they're offset
and remain consistently offset from their creation.